arm: semihosting: print console output on STDOUT instead of STDERR
Required if we use semihosting as default output for ARM BB targets.

Part of P808-009.

--- merged with ---

ARM Semihosting: Add -semihosting-console option

to be able to define a chardev for the semihosting (file, tcp, etc.).

If the option is not used, the semihosting console will print in the
first serial chardev, if any, or STDERR.

--- merged with ---

arm-semi: fix the output to chardev

This fixes the output to chardev.

2019-02-28  KONRAD Frederic  <konrad@adacore.com>
        * vl.c: added qemu_chr_new argument.

TN: Q705-031
Change-Id: Ibc4ff8d574e0e17894754cc1e4f16c9b49c08266